+++ src/OneDSchrodingerEquationSolver.jl
@@ -6,10 +6,10 @@ using ..IntegralNumericalMethods
 using ..MathUtils
 
 raw"""
-    solver(E::Float32,init_valu1_fwrd::Float32,
-    init_valu2_fwrd::Float32, init_valu1_bwrd::Float32,
-    init_valu2_bwrd::Float32, v_effe::Vector{Float32},
-    grid::Vector{Float32})::Tuple{Vector{Float32},Float32}
+    solver(E::Float64,init_valu1_fwrd::Float64,
+    init_valu2_fwrd::Float64, init_valu1_bwrd::Float64,
+    init_valu2_bwrd::Float64, v_effe::Vector{Float64},
+    grid::Vector{Float64})::Tuple{Vector{Float64},Float64}
 
     Returns the normalized wave function (u(r)) of one dimensional shcrodinger equation of the form
     \ frac{-1}{2} \ frac{d^2 u(r)}{dr^2} + V_{effe} u(r) = E u(r) 
@@ -18,22 +18,22 @@ raw"""
     It also returns the merging value to help in the finding of the energy eigenvalue.
 
 **Inputs:**
-- `E::Float32`: Value to use as energy eigenvalue
-- `init_valu1_fwrd::Float32`: Value of u(grid[1]) to use as first initial value in forward solution.
-- `init_valu2_fwrd::Float32`: Value of u(grid[2]) to use as second initial value in forward solution.
-- `init_valu1_bwrd::Float32`: Value of u(grid[end]) to use as first initial value in backward solution.
-- `init_valu2_bwrd::Float32`: Value of u(grid[end -1]) to use as second initial value in backward solution.
-- `v_effe::Vector{Float32}`: Vector with the values of the effective potential over the grid.
-- `grid::Vector{Float32}`: Vector with the grid values.
+- `E::Float64`: Value to use as energy eigenvalue
+- `init_valu1_fwrd::Float64`: Value of u(grid[1]) to use as first initial value in forward solution.
+- `init_valu2_fwrd::Float64`: Value of u(grid[2]) to use as second initial value in forward solution.
+- `init_valu1_bwrd::Float64`: Value of u(grid[end]) to use as first initial value in backward solution.
+- `init_valu2_bwrd::Float64`: Value of u(grid[end -1]) to use as second initial value in backward solution.
+- `v_effe::Vector{Float64}`: Vector with the values of the effective potential over the grid.
+- `grid::Vector{Float64}`: Vector with the grid values.
 """
 
-function solver(E::Float32,init_valu1_fwrd::Float32,
-    init_valu2_fwrd::Float32, init_valu1_bwrd::Float32,
-    init_valu2_bwrd::Float32, v_effe::Vector{Float32},
-    grid::Vector{Float32})::Tuple{Vector{Float32},Float32}
+function solver(E::Float64,init_valu1_fwrd::Float64,
+    init_valu2_fwrd::Float64, init_valu1_bwrd::Float64,
+    init_valu2_bwrd::Float64, v_effe::Vector{Float64},
+    grid::Vector{Float64})::Tuple{Vector{Float64},Float64}
 
-    f::Vector{Float32}= 2.0.*(v_effe .- E);
-    g=zeros(Float32, size(f)[1]);
+    f::Vector{Float64}= 2.0.*(v_effe .- E);
+    g=zeros(Float64, size(f)[1]);
 
     #find turn_pnts of of f, basically the clasical turning points of the effective density_potential
     #with restepect to the E proposed eigenvalue
