+++ test/uS1HydrogenTest.jl
@@ -8,22 +8,22 @@ import .SchrodingerEquationSolver.IntegralNumericalMethods as IntegralNumericalM
 import .SchrodingerEquationSolver.MathUtils as MathUtils
 @testset "uS1HydrogenTest" begin
         
-        r_max::Float32=10.0;
-        l::Int32=0;
-        Z::Int32=1;
-        E::Float32= -0.5000;
+        r_max::Float64=10.0;
+        l::Int64=0;
+        Z::Int64=1;
+        E::Float64= -0.5000;
 
         grid= Grids.exponential_grid(r_max, Z);
         v_colu= Potentials.coulomb_potential(Z, grid);
         u_s1_hydr= Hydrogen.u_s1_hydrogen(grid);
-        init_valu1_fwrd::Float32=u_s1_hydr[1];#grid[1]^(l+1.0)
-        init_valu2_fwrd::Float32=u_s1_hydr[2];#grid[2]^(l+1.0)
-        init_valu1_bwrd::Float32=u_s1_hydr[end];#grid[1]^(l+1.0)
-        init_valu2_bwrd::Float32=u_s1_hydr[end-1];#grid[2]^(l+1.0)
+        init_valu1_fwrd::Float64=u_s1_hydr[1];#grid[1]^(l+1.0)
+        init_valu2_fwrd::Float64=u_s1_hydr[2];#grid[2]^(l+1.0)
+        init_valu1_bwrd::Float64=u_s1_hydr[end];#grid[1]^(l+1.0)
+        init_valu2_bwrd::Float64=u_s1_hydr[end-1];#grid[2]^(l+1.0)
         v_effe= v_colu;
-        f::Vector{Float32}= 2.0.*(v_effe .- E);
+        f::Vector{Float64}= 2.0.*(v_effe .- E);
         #given f find the classical turning points
-        g=zeros(Float32, size(f)[1]);
+        g=zeros(Float64, size(f)[1]);
         turn_pnts= MathUtils.indices_of_zeros_finder(f);
 
         #do forward integration of radial shcrodinger equation u
